Fix auto-resizing of BBCode editor text box on Firefox
The previous implementation strategy used getComputedStyle().cssText to
copy styles from the real textarea onto the fake, invisible one used for
computing sizing metrics. This worked alright on Chrome, but not on
Firefox, as Firefox currently always returns an empty string for cssText
on the result of getComputedStyle() (see Bugzilla #137687).

This commit avoids copying the computed styles manually in favor of just
applying the same CSS classes to the invisible textarea. It also tweaks
the size computation to add the border height to the computed
scrollHeight before setting the inline height style, since scrollHeight
does not include the border, but box-sizing is set to border-box, and
the 2px difference was enough to provoke a scroll bar in Firefox.